chore: split builder and visitor
